==> fgwas/diagram_hrc/cross_tissue/03.1_enrichment-plots.Rmd <==
---
title: "03_enrichment-plots"
author: "Jason Torres"
date: "September 2017"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Setup libraries and directories

```{r}
"%&%" <- function(a,b) paste0(a,b)
library("dplyr")
library("data.table")
library("ggplot2")
library("gridExtra")

serv.dir <- "/Users/jtorres/FUSE/"
fgwas.dir <- serv.dir %&% "projects/t2d-integration/fgwas/diagram_hrc/cross_tissue/"
plot.dir <- fgwas.dir %&% "plots/"

output.prefix <- "DIAMANTE-UKBB-gwas"

annotation.keyfile <- NULL #fgwas.dir %&% "plot-key.csv"# Optional: Specifiy if you have a csv-delimited txt file with annotation names (as they appear in the fgwas input file) on the left column and plotting names on the right column, column names should be Name and PlotName, otherwise set to NULL
if (!is.null(annotation.keyfile)){
  key.df <- fread(annotation.keyfile,sep=",",sep2=" ")
}



```

# Processing function

```{r}

annot_process <- function(fname){
  df <- fread(fname)
  df$CI_lo <- gsub("<","",df$CI_lo); df$CI_lo <- gsub("fail",NA,df$CI_lo); df$CI_lo <- as.numeric(df$CI_lo)
  df$CI_hi <- gsub(">","",df$CI_hi); df$CI_hi <- gsub("fail",NA,df$CI_hi); df$CI_hi <- as.numeric(df$CI_hi)
  return(df)
}

```

# Build data frames for plotting  

```{r}


update_names <- function(par.df, key.df){
  par.df$parameter <- gsub("_ln","",par.df$parameter)
  for (i in 1:dim(par.df)[1]){
    annot <- par.df$parameter[i]
    if (grepl("distance_tss",annot)==TRUE){
      annot <- "distance_tss"
    }
    plotname <- key.df$PlotName[grepl(annot,key.df$Name)]
    par.df$parameter[i] <- plotname
    print(c(annot,plotname))
  }
  return(par.df)
}

build_param_df <- function(fgwas.out.dir,name.vec=NULL){
  files <- list.files(fgwas.out.dir)
  param.files <- files[grepl(".params",files,fixed=TRUE)]
  param.files <- param.files[!grepl("+",param.files,fixed=TRUE)]
  param.files <- param.files[!grepl("drop",param.files,fixed=TRUE)]  
  param.files <- param.files[!grepl("best-joint-model",param.files,fixed=TRUE)]  

  out.df <- c()
  for (f in param.files){
    stack.df <- annot_process(fgwas.out.dir%&%f)
    out.df <- rbind(out.df,stack.df)
  }
  out.df <- filter(out.df,parameter!="pi_region")
  out.df <- arrange(out.df,desc(estimate))
  enriched <- filter(out.df, CI_lo>0,estimate>0,CI_hi>0)
  depleted <- filter(out.df, CI_lo<0,estimate<0,CI_hi<0)
  print("There are " %&% dim(out.df)[1] %&% " annotations")
  print("There are " %&% dim(enriched)[1] %&% " annotations that are marginally ENRICHED")
  print("There are " %&% dim(depleted)[1] %&% " annotations that are marginally DEPLETED")
  out.df <- rbind(enriched,depleted)
  #if (!is.null(name.vec)){
  #  out.df$parameter<-name.vec
  #}
  if (!is.null(annotation.keyfile)){
    out.df <- update_names(out.df,key.df)
  }
  return(out.df)
}



build_param_best_df <- function(pre){
  # Note: pre is best model from fgwas run 
  param.df <- annot_process(pre %&% ".params")
  param.df <- arrange(param.df,desc(estimate))
  param.df <- filter(param.df,parameter!="pi")
  param.df <- filter(param.df,parameter!="pi_region")

  #if (!is.null(name.vec)){
  #  param.df$parameter<-name.vec
  #}
  if (!is.null(annotation.keyfile)){
    param.df <- update_names(param.df,key.df)
  }
  return(param.df)
}


```


# Annotation plot

```{r}


an_plot <- function(mydf,mytitl="",mylow=-10,myhigh=10,interval=1){
  mydf <- within(mydf,parameter<-factor(mydf$parameter,levels=rev(mydf$parameter)))
  mydf <- filter(mydf,parameter!="pi_region")
  plt <- ggplot(data=mydf,aes(x=parameter,y=estimate)) +
      geom_hline(yintercept=0,linetype=2) +
    ylab("Log2FE") + xlab("Annotation") +
    geom_errorbar(aes(ymin=CI_lo,ymax=CI_hi),width=0.1) +
    geom_point(shape=21,size=1.5,col="black",aes(fill=(estimate>0)))  +
    scale_fill_manual(values=c("steelblue1","firebrick1")) +
    theme_bw()  +  theme(legend.position = "none",
                         panel.grid.minor=element_blank(),
                         panel.grid.major=element_blank()) +
    scale_y_continuous(breaks=seq(mylow,myhigh,by=interval)) +
    coord_flip(ylim=c(mylow,myhigh)) +
    ggtitle(mytitl)
  return(plt)
}


```


# Round 1 Plots 

```{r}

plot_generator <- function(tissue,round=1,key=FALSE){
  # Seperate  
  if (round==1){
    fgwas.out.dir <- fgwas.dir %&% "fgwas_output_" %&% tissue %&% "/"
    my.title <- tissue %&% ": " %&% "ChIP-seq only"
  } else if (round==2){
    fgwas.out.dir <- fgwas.dir %&% "fgwas_output_" %&% tissue %&% "/round2/"
    my.title <- tissue %&% ": " %&% "ChIP-seq + ATAC"
    if (key==TRUE){
      fgwas.in.dir <- fgwas.dir %&% "fgwas_input_" %&% tissue %&% "/round2/"
      key.df <- fread(fgwas.in.dir%&%"annotation_key-file.txt")
  }
  } else{
    stop("Must be round 1 or 2")
  }
  best.joint.model.file.prefix <- fgwas.out.dir %&% "best-joint-model"
  param.sing.df <- build_param_df(fgwas.out.dir)
  if (key==TRUE){
    param.sing.df$parameter <- sapply(1:length(param.sing.df$parameter),function(i){
      k <- param.sing.df$parameter[i]
      k <- gsub("_ln","",k); k <- gsub("_0_5000","",k)
      a <- filter(key.df,Key==k)$Annotation
      return(a)
    })
  }
  
  plt1 <- an_plot(param.sing.df,mytitl=my.title%&%": Seperate Analysis",
                  mylow=-5,myhigh=5,interval=1)
  #plt1
  write.table(x=param.sing.df,file=plot.dir %&% "round" %&% round %&% 
           "_" %&% tissue %&% ".fgwas_enrich.seperate.txt",
              sep="\t",quote=F,col.names=TRUE,row.names=FALSE)
  ggsave(filename=plot.dir %&% "round" %&% round %&% 
           "_" %&% tissue %&% ".fgwas_enrich.seperate.png",plot=plt1,
         width=5,height=5)  
  # Joint 
  param.df <- build_param_best_df(pre=best.joint.model.file.prefix)
  if (key==TRUE){
    param.df$parameter <- sapply(1:length(param.df$parameter),function(i){
      k <- param.df$parameter[i]
      k <- gsub("_ln","",k); k <- gsub("_0_5000","",k)
      a <- filter(key.df,Key==k)$Annotation
      return(a)
    })
  }
  plt2 <- an_plot(param.df,mytitl=my.title%&%": Joint Analysis",
                  mylow=-5,myhigh=5,interval=1)
  #plt2
  write.table(x=param.df,file=plot.dir %&% "round" %&% round %&% 
           "_" %&% tissue %&% ".fgwas_enrich.joint.txt",
              sep="\t",quote=F,col.names=TRUE,row.names=FALSE)
  ggsave(filename=plot.dir %&% "round" %&% round %&% 
           "_" %&% tissue %&% ".fgwas_enrich.joint.png",plot=plt2,
         width=5,height=5)  
}


```




```{r}

tiss.vec <- c("islet","liver","adipose","muscle")
tiss.vec <- c("adipose")

for (tiss in tiss.vec){
  #plot_generator(tiss,round=1)
  plot_generator(tiss,round=2,key=TRUE)
}


```



# Tissue Comparator plots 


```{r}


build_param_all_df <- function(fgwas.out.dir,name.vec=NULL){
  files <- list.files(fgwas.out.dir)
  param.files <- files[grepl(".params",files,fixed=TRUE)]
  param.files <- param.files[!grepl("+",param.files,fixed=TRUE)]
  param.files <- param.files[!grepl("drop",param.files,fixed=TRUE)]  
  param.files <- param.files[!grepl("best-joint-model",param.files,fixed=TRUE)]  
  param.files <- param.files[!grepl("SINGLE_fgwas_run_loci-partition",param.files,fixed=TRUE)]  

  out.df <- c()
  for (f in param.files){
    stack.df <- annot_process(fgwas.out.dir%&%f)
    out.df <- rbind(out.df,stack.df)
  }
  out.df <- filter(out.df,parameter!="pi_region")
  out.df <- arrange(out.df,desc(estimate))
  enriched <- filter(out.df, CI_lo>0,estimate>0,CI_hi>0)
  depleted <- filter(out.df, CI_lo<0,estimate<0,CI_hi<0)
  print("There are " %&% dim(out.df)[1] %&% " annotations")
  print("There are " %&% dim(enriched)[1] %&% " annotations that are marginally ENRICHED")
  print("There are " %&% dim(depleted)[1] %&% " annotations that are marginally DEPLETED")
  #out.df <- rbind(enriched,depleted)
  #if (!is.null(name.vec)){
  #  out.df$parameter<-name.vec
  #}
  if (!is.null(annotation.keyfile)){
    out.df <- update_names(out.df,key.df)
  }
  return(out.df)
}


get_tiss_df <- function(tissue,round=1){
  # Seperate  
  if (round==1){
    fgwas.out.dir <- fgwas.dir %&% "fgwas_output_" %&% tissue %&% "/"
    my.title <- tissue %&% ": " %&% "ChIP-seq only"
  } else if (round==2){
    fgwas.out.dir <- fgwas.dir %&% "fgwas_output_" %&% tissue %&% "/round2/"
    my.title <- tissue %&% ": " %&% "ChIP-seq + ATAC"
    if (key==TRUE){
      fgwas.in.dir <- fgwas.dir %&% "fgwas_input_" %&% tissue %&% "/round2/"
      key.df <- fread(fgwas.in.dir%&%"annotation_key-file.txt")
  }
  } else{
    stop("Must be round 1 or 2")
  }
  param.sing.df <- build_param_all_df(fgwas.out.dir)  
  return(param.sing.df)
}

tissue_compare_plots <- function(round=1){
  islet.df <- get_tiss_df("islet") %>% filter(.,grepl("Active_enhancer",parameter)); islet.df$parameter <- "islet_" %&% islet.df$parameter
  liver.df <- get_tiss_df("liver") %>% filter(.,grepl("Active_enhancer",parameter)); liver.df$parameter <- "liver_" %&% liver.df$parameter
  muscle.df <- get_tiss_df("muscle") %>% filter(.,grepl("Active_enhancer",parameter)); muscle.df$parameter <- "muscle_" %&% muscle.df$parameter
  adipose.df <- get_tiss_df("adipose") %>% filter(.,grepl("Active_enhancer",parameter)); adipose.df$parameter <- "adipose_" %&% adipose.df$parameter
  enhancer.df <- rbind(islet.df,liver.df,muscle.df,adipose.df)
  enh.plt <- an_plot(enhancer.df,mytitl="Enhancers"%&%": Seperate Analysis",
                   mylow=-5,myhigh=5,interval=1)
  islet.df <- get_tiss_df("islet") %>% filter(.,grepl("Weak_enhancer",parameter)); islet.df$parameter <- "islet_" %&% islet.df$parameter
  liver.df <- get_tiss_df("liver") %>% filter(.,grepl("Weak_enhancer",parameter)); liver.df$parameter <- "liver_" %&% liver.df$parameter
  muscle.df <- get_tiss_df("muscle") %>% filter(.,grepl("Weak_enhancer",parameter)); muscle.df$parameter <- "muscle_" %&% muscle.df$parameter
  adipose.df <- get_tiss_df("adipose") %>% filter(.,grepl("Weak_enhancer",parameter)); adipose.df$parameter <- "adipose_" %&% adipose.df$parameter
  enhancer.df <- rbind(islet.df,liver.df,muscle.df,adipose.df)
  enhweak.plt <- an_plot(enhancer.df,mytitl="Weak Enhancers"%&%": Seperate Analysis",
                   mylow=-5,myhigh=5,interval=1)
  islet.df <- get_tiss_df("islet") %>% filter(.,grepl("Active_TSS",parameter)); islet.df$parameter <- "islet_" %&% islet.df$parameter
  liver.df <- get_tiss_df("liver") %>% filter(.,grepl("Active_TSS",parameter)); liver.df$parameter <- "liver_" %&% liver.df$parameter
  muscle.df <- get_tiss_df("muscle") %>% filter(.,grepl("Active_TSS",parameter)); muscle.df$parameter <- "muscle_" %&% muscle.df$parameter
  adipose.df <- get_tiss_df("adipose") %>% filter(.,grepl("Active_TSS",parameter)); adipose.df$parameter <- "adipose_" %&% adipose.df$parameter
  tss.df <- rbind(islet.df,liver.df,muscle.df,adipose.df)
  tss.plt <- an_plot(tss.df,mytitl="TSS"%&%": Seperate Analysis",
                   mylow=-5,myhigh=5,interval=1) 
  islet.df <- get_tiss_df("islet") %>% filter(.,grepl("Weak_TSS",parameter)); islet.df$parameter <- "islet_" %&% islet.df$parameter
  liver.df <- get_tiss_df("liver") %>% filter(.,grepl("Weak_TSS",parameter)); liver.df$parameter <- "liver_" %&% liver.df$parameter
  muscle.df <- get_tiss_df("muscle") %>% filter(.,grepl("Weak_TSS",parameter)); muscle.df$parameter <- "muscle_" %&% muscle.df$parameter
  adipose.df <- get_tiss_df("adipose") %>% filter(.,grepl("Weak_TSS",parameter)); adipose.df$parameter <- "adipose_" %&% adipose.df$parameter
  tssweak.df <- rbind(islet.df,liver.df,muscle.df,adipose.df)
  tssweak.plt <- an_plot(tssweak.df,mytitl="Weak TSS"%&%": Seperate Analysis",
                   mylow=-5,myhigh=5,interval=1)   
  grid.arrange(enh.plt,tss.plt,enhweak.plt,tssweak.plt,ncol=2,nrow=2)
  c(fread(fgwas.dir %&% "fgwas_output_" %&% "islet" %&% "/" %&% "10_Active_enhancer_2.llk")[1,]$V2,
    fread(fgwas.dir %&% "fgwas_output_" %&% "liver" %&% "/" %&% "10_Active_enhancer_2.llk")[1,]$V2,
    fread(fgwas.dir %&% "fgwas_output_" %&% "muscle" %&% "/" %&% "10_Active_enhancer_2.llk")[1,]$V2,
    fread(fgwas.dir %&% "fgwas_output_" %&% "adipose" %&% "/" %&% "10_Active_enhancer_2.llk")[1,]$V2)
  c(fread(fgwas.dir %&% "fgwas_output_" %&% "islet" %&% "/" %&% "1_Active_TSS.llk")[1,]$V2,
    fread(fgwas.dir %&% "fgwas_output_" %&% "liver" %&% "/" %&% "1_Active_TSS.llk")[1,]$V2,
    fread(fgwas.dir %&% "fgwas_output_" %&% "muscle" %&% "/" %&% "1_Active_TSS.llk")[1,]$V2,
    fread(fgwas.dir %&% "fgwas_output_" %&% "adipose" %&% "/" %&% "1_Active_TSS.llk")[1,]$V2)


}

```
