==> fgwas/diagram_hrc/cross_tissue/04.1_get-sig-blocks.Rmd <==
---
title: "functional_credible_sets.Rmd"
author: "Jason Torres"
date: "May 22, 2017"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


Constructing functional (fgwas) credible sets from the fgwas model with the best cross validated likelihood 

# Setup 

```{r}

"%&%" <- function(a,b) paste0(a,b)

library("data.table")
library("dplyr")
library("ggplot2")
library("GenomicRanges")

serv.dir <- "/Users/jtorres/FUSE/"

work.dir <- serv.dir %&% "projects/t2d-integration/fgwas/diagram_hrc/cross_tissue/"
cred.set.dir <- serv.dir %&% "projects/t2d-integration/fgwas/diagram_hrc/cross_tissue/credible_sets/"

```


```{r}

get_sig_blocks <- function(pre, tissue, round, ppa.thresh=0.90){
  blk.df <- fread("cat " %&% pre %&% ".segbfs.gz" %&% " | zmore")
  sig.blk.df <- filter(blk.df, PPA >= ppa.thresh) %>% arrange(desc(PPA))
  chrom <- as.integer(gsub("chr","",sig.blk.df$chr))
  sig.blk.df$chrom <- chrom
  sig.blk.df <- arrange(sig.blk.df,chrom,st) %>%
    dplyr::select(one_of("chr","st","sp","PPA"))
  names(sig.blk.df) <- c("CHR","blk.start","blk.end","blk.PPA")
  SEGNUMBER <- 1:dim(sig.blk.df)[1]
  sig.blk.df <- cbind(SEGNUMBER,sig.blk.df)
  write.table(x=sig.blk.df,file = cred.set.dir %&% "Round" %&% 
                round  %&% "." %&% tissue %&% ".sig_blocks_ppa"%&%ppa.thresh%&%".txt", 
              sep="\t",quote=FALSE,row.names=FALSE)
  prop.sig <- dim(sig.blk.df)[1]/dim(blk.df)[1] # 0.0582762
  print("Proportion of blocks with cummulative PPA >= " %&% ppa.thresh %&% ": " %&% prop.sig)
  return(sig.blk.df)
}

```


Test 

```{r}

round1 <- function(tiss.vec=c("islet","liver","adipose","muscle")){
  print("Round 1")
  for (tiss in tiss.vec){
    fgwas.output.dir <- work.dir %&% "fgwas_output_" %&% tiss %&% "/"
    pre <- fgwas.output.dir %&% "best-joint-model"  
    get_sig_blocks(pre,tiss,1) 
  }  
}

round2 <- function(tiss.vec=c("liver","adipose","muscle")){
  print("Round 2")
  for (tiss in tiss.vec){
    fgwas.output.dir <- work.dir %&% "fgwas_output_" %&% tiss %&% "/round2/"
    pre <- fgwas.output.dir %&% "best-joint-model"  
    get_sig_blocks(pre,tiss,2) 
  }  
}

round1()
round2()
```



# Subset fgwas SNP output file to SNP in significant blocks 


```{bash}

python subset_to_seg.py pre fgwas_blk-t2d_loci_152.txt

```
