==> fgwas/diagram_hrc/cross_tissue/05.1_create-best-input-bed.Rmd <==
---
title: "05.1"
author: "Jason Torres"
date: "November 2017"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Setup libraries and directories

```{r}

"%&%" <- function(a,b) paste0(a,b)
library("data.table")
library("tidyverse")

serv.dir <- "/Users/jtorres/FUSE/"
fgwas.dir <- serv.dir %&% "projects/t2d-integration/fgwas/diagram_hrc/cross_tissue/"

best.joint.model.file.prefix <- "best-joint-model"


```

# Create and write input file


```{r}

write_input_bed <- function(tissue){
  
  output.dir <- fgwas.dir %&% "conditional/fgwas_input_files_" %&% tissue %&% "/"
  dir.create(output.dir)

  
  fgwas.in.dir <- fgwas.dir %&% "fgwas_input_" %&% tissue %&% "/"
  fgwas.out.dir <- fgwas.dir %&% "fgwas_output_" %&% tissue %&% "/"
  bed.df <- fread(fgwas.in.dir %&% "anno_input.bed")
  bed.df$V4 <- gsub("/","_",bed.df$V4,fixed=TRUE)
  #key.df <- fread(fgwas.in.dir %&% "annotation_key-file.txt")
  best.df <- fread(fgwas.out.dir %&% best.joint.model.file.prefix %&% ".params") %>% filter(parameter!="pi_region")
  best.df$parameter <- gsub("_ln","",best.df$parameter) %>% gsub("_0_5000","",.)
  #best.df$parameter <- as.character(map(best.df$parameter,function(x){filter(key.df,Key==x)$Annotation}))
  
  keep.df <- filter(bed.df, V4 %in% best.df$parameter)
  write.table(x=keep.df,file=output.dir %&% "best_anno_input.bed",sep="\t",row.names=F,col.names=F,quote=F)
}


```


Write files

```{r}

tiss.vec <- c("adipose","muscle","liver","islet")
for (tiss in tiss.vec){
  print(tiss)
  write_input_bed(tiss)
}


```
