==> fgwas/diagram_hrc/cross_tissue/07.2_build-analysis-matrices.Rmd <==
---
title: "Create analysis matrices"
author: "Jason Torres"
date: "11/20/2017"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


# Setup 

```{r}

"%&%" <- function(a,b) paste0(a,b)
library("data.table")
library("tidyverse")
library("GenomicRanges")

serv.dir <- "/Users/jtorres/FUSE/" 

work.dir <- serv.dir %&% "projects/t2d-integration/fgwas/diagram_hrc/cross_tissue/"
cred.dir <- serv.dir %&% "projects/t2d-integration/fgwas/diagram_hrc/cross_tissue/credible_sets/"
write.dir <- work.dir %&% "analysis_files/"

ignore.names <- c("Locus.ID","SEGNUMBER","SNPID","CHR","POS","logBF","Z","PPA","symbol")

```


Build credible set dataframes 


# Functions 

Get ppa annotation matrix 

```{r}

get_full_df <- function(cred.df){
  loc.vec <- unique(cred.df$Locus.ID)
  annot.vec <- names(cred.df)[!(names(cred.df) %in% ignore.names)]
  
  out.df <- c() 
  print(length(loc.vec))
  pb <- txtProgressBar(min=0,max=length(loc.vec),style=3)
  for (i in 1:length(loc.vec)){
    setTxtProgressBar(pb,i)
    loc <- loc.vec[i]
    temp.df <- filter(cred.df,Locus.ID==loc)
    cumppa <- sum(temp.df$PPA)
    myrow <- c(loc)
    for (annot in annot.vec){
      index <- match(annot,names(temp.df))
      #sub.df <- temp.df[,c(1:9,index),]
      sub.df <- temp.df %>% dplyr::select(.,one_of(ignore.names,annot))
      n <- length(names(sub.df))
      annotppa <- sum(sub.df$PPA[as.logical(sub.df[,n])]) / cumppa
      myrow <- c(myrow,annotppa)
    }
    other <- (1-sum(unique(as.numeric(myrow[2:(length(myrow))]))))
    other <- ifelse(other>=0,other,0)
    myrow <- c(myrow,other)
    cumsum <- sum((as.numeric(myrow[2:(length(myrow))])))
    myrow[2:length(myrow)] <- as.numeric(myrow[2:length(myrow)]) / cumsum # scale to give proper proportion
    names(myrow) <- c("Locus.ID",annot.vec,"other")
    myrow <- as.data.frame(t(as.matrix(myrow)))
    out.df <- rbind(out.df,myrow)
  }
  out.df <- as.data.frame(out.df)
  row.names(out.df) <- NULL
  indx <- sapply(out.df, is.factor)
  out.df[indx] <- lapply(out.df[indx], function(x) as.numeric(as.character(x)))
  out.df$Locus.ID <- loc.vec
  return(out.df)
}

build_ppa_matrix <- function(tissue){
  block99.df <- fread(cred.dir %&% tissue %&% "/" %&% "results_blocks.txt",sep="\t")
  fcred99.df <- fread(cred.dir %&% tissue %&% "/" %&% "results_func-cred-sets.txt",sep="\t")
  full.df <- get_full_df(fcred99.df)
  write.table(x=full.df,file=write.dir%&%tissue %&% "_best-joint_ppa-prop.txt",
              sep="\t",quote=FALSE,row.names=FALSE)
}

tiss.vec <- c("islet","muscle","liver","adipose")

for (tiss in tiss.vec){
  print(tiss)
  build_ppa_matrix(tiss)
}


```



# Build master analysis data frame for tissue-classification project

```{r}

null.dir <- serv.dir %&% "projects/t2d-integration/fgwas/diagram_hrc/ukbb-diamante-euro-manuscript/null/credible_sets/"
nullblock.df <- fread(null.dir %&% "null_results_blocks.txt",sep="\t")
nullfcred.df <- fread(null.dir %&% "null_results_func-cred-sets.txt",sep="\t")

build_classify_df <- function(tissue){
  block99.df <- fread(cred.dir %&% tissue %&% "/" %&% "results_blocks.txt",sep="\t")
  fcred99.df <- fread(cred.dir %&% tissue %&% "/" %&% "results_func-cred-sets.txt",sep="\t")
  loc.vec <- unique(fcred99.df$Locus.ID)
  block99.df <- filter(block99.df,Locus.ID%in%loc.vec)
  out.df <- block99.df %>% dplyr::select(.,one_of("Locus.ID","loc.pos","CHR","refseq"))
  out.df$null.num <- map(out.df$Locus.ID,function(id){
    filter(nullfcred.df,Locus.ID==id)$PPA %>% length(.)
  }) %>% as.numeric(.)
  out.df$null.len <- map(out.df$Locus.ID,function(id){
    filter(nullblock.df,Locus.ID==id)$fcred.len 
  }) %>% as.numeric(.)
  out.df$maxpp <- map(out.df$Locus.ID,function(id){
    filter(fcred99.df,Locus.ID==id)$PPA %>% max(.)
  }) %>% as.numeric(.)
  out.df$maxppDelta <- map(out.df$Locus.ID,function(id){
    m0 <- filter(nullfcred.df,Locus.ID==id)$PPA %>% max(.)
    m1 <- filter(fcred99.df,Locus.ID==id)$PPA %>% max(.)
    return((m1-m0)/m0)
  }) %>% as.numeric(.)
  out.df$numDelta <- map(out.df$Locus.ID,function(id){
    m0 <- filter(nullfcred.df,Locus.ID==id)$PPA %>% length(.)
    m1 <- filter(fcred99.df,Locus.ID==id)$PPA %>% length(.)
    return((m1-m0)/m0)
  }) %>% as.numeric(.)
  out.df$lenDelta <- map(out.df$Locus.ID,function(id){
    m0 <- filter(nullblock.df,Locus.ID==id)$fcred.len 
    m1 <- filter(block99.df,Locus.ID==id)$fcred.len 
    if (m0!=0){
      return((m1-m0)/m0)
    } else{
      return(0)
    }
  }) %>% as.numeric(.)
  names(out.df)[(length(names(out.df))-3):length(names(out.df))] <- tissue %&% "." %&% 
    names(out.df)[(length(names(out.df))-3):length(names(out.df))] 
  return(out.df)
}

build_classify_df_full <- function(tiss.vec=c("islet","muscle","liver","adipose")){
  print(tiss.vec[1])
  out.df <- build_classify_df(tiss.vec[1])
  for (tiss in tiss.vec[2:length(tiss.vec)]){
    print(tiss)
    build.df <- build_classify_df(tiss) %>% dplyr::select(.,contains(tiss))
    out.df <- cbind(out.df,build.df)
  }
  return(out.df)
}

classify.df <- build_classify_df_full()
write.table(x=classify.df,file=write.dir%&%"loci_profile.txt",
            sep="\t",quote=FALSE,row.names=FALSE)

```
